requirements.txt
lightrag-hku
PyPDF2
copy
functools
json
aioboto3
aiohttp
numpy
ollama
openai
base64
struct
tenacity
transformers
torch
pydantic
typing
